Add event-repository with curl-url request
