+++ IOS@Taipei/IOS@Taipei/UI/FirstViewController.swift
@@ -13,8 +13,7 @@ class FirstViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
 
-        FacebookEventRepository.shared.setup(with: Secrets.accessToken.rawValue)
-        FacebookEventRepository.shared.fetchFacebookEvents { (result) in
+        EventRepository.shared.fetchFacebookEvents { (result) in
             switch result {
             case .success(let events):
                 print("events -> \(events)")
@@ -23,7 +22,6 @@ class FirstViewController: UIViewController {
             }
         }
     }
-
 
 }
